backend-python/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
google-generativeai
python-dotenv
requests